# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EventFlow Virtual Event Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus];

export const EventStatus = {
  draft: "draft",
  upcoming: "upcoming",
  live: "live",
  ended: "ended",
  cancelled: "cancelled",
} as const;

export interface Event {
  id: number;
  title: string;
  description: string;
  hostName: string;
  scheduledAt: string;
  endsAt?: string;
  status: EventStatus;
  category: string;
  maxAttendees?: number | null;
  registrationCount: number;
  streamUrl?: string | null;
  thumbnailUrl?: string | null;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateEventBody {
  title: string;
  description: string;
  hostName: string;
  scheduledAt: string;
  endsAt?: string | null;
  category: string;
  maxAttendees?: number | null;
  streamUrl?: string | null;
  thumbnailUrl?: string | null;
  isPublic?: boolean;
}

export type UpdateEventBodyStatus =
  (typeof UpdateEventBodyStatus)[keyof typeof UpdateEventBodyStatus];

export const UpdateEventBodyStatus = {
  draft: "draft",
  upcoming: "upcoming",
  live: "live",
  ended: "ended",
  cancelled: "cancelled",
} as const;

export interface UpdateEventBody {
  title?: string;
  description?: string;
  hostName?: string;
  scheduledAt?: string;
  endsAt?: string | null;
  category?: string;
  maxAttendees?: number | null;
  streamUrl?: string | null;
  thumbnailUrl?: string | null;
  isPublic?: boolean;
  status?: UpdateEventBodyStatus;
}

export interface EventListResponse {
  events: Event[];
  total: number;
}

export type RegistrationStatus =
  (typeof RegistrationStatus)[keyof typeof RegistrationStatus];

export const RegistrationStatus = {
  confirmed: "confirmed",
  waitlisted: "waitlisted",
  cancelled: "cancelled",
} as const;

export interface Registration {
  id: number;
  eventId: number;
  attendeeName: string;
  attendeeEmail: string;
  status: RegistrationStatus;
  registeredAt: string;
  event?: Event | null;
}

export interface RegisterAttendeeBody {
  attendeeName: string;
  attendeeEmail: string;
}

export interface RegistrationListResponse {
  registrations: Registration[];
  total: number;
}

export interface ChatMessage {
  id: number;
  eventId: number;
  senderName: string;
  message: string;
  isPinned: boolean;
  isDeleted: boolean;
  sentAt: string;
}

export interface SendChatMessageBody {
  senderName: string;
  message: string;
}

export interface ChatMessageListResponse {
  messages: ChatMessage[];
  total: number;
}

export type StreamStatusStatus =
  (typeof StreamStatusStatus)[keyof typeof StreamStatusStatus];

export const StreamStatusStatus = {
  live: "live",
  ended: "ended",
} as const;

export interface StreamStatus {
  eventId: number;
  status: StreamStatusStatus;
  streamUrl?: string | null;
  startedAt?: string | null;
}

export interface DashboardStats {
  totalEvents: number;
  liveEvents: number;
  upcomingEvents: number;
  totalRegistrations: number;
  totalAttendees: number;
  eventsThisMonth: number;
  registrationsThisMonth: number;
}

export type ListEventsParams = {
  status?: ListEventsStatus;
  search?: string;
  limit?: number;
  offset?: number;
};

export type ListEventsStatus =
  (typeof ListEventsStatus)[keyof typeof ListEventsStatus];

export const ListEventsStatus = {
  upcoming: "upcoming",
  live: "live",
  ended: "ended",
  all: "all",
} as const;

export type ListChatMessagesParams = {
  limit?: number;
  before?: number;
};